Update GuiManager_[v1].lua
Seeing as this is the first released version, there isn’t tons of functionality but there’s enough to get some decent use out of it, like dragging and rotating (which have events associated with them). Documentation can be found at the bottom of the module, I may make a GitHub Pages website to put more detailed documentation on.